sunucu-todo/requirements.txt
argon2-cffi==23.1.0
fastapi==0.115.12
httpx==0.28.1
itsdangerous==2.2.0
pytest==8.3.5
uvicorn==0.34.0
